Moved opcode accessors to Defines.h.
This allows me to centralize the immediate feature accessors. This is a new feature that uses the upper 2 bits of the Opcode to as an additional operand for some Opcodes. For instance RETI will be a single byte opcode that uses the upper 2 bits as the return count, 0..3.

+++ src/Containers.h
@@ -397,7 +397,7 @@ public:
     }
 
     bool empty() const { return _size == 0; }
-    size_t size() const { return _size; };
+    uint32_t size() const { return _size; };
     const T& operator[](uint16_t i) const { return at(i); };
     
     const T& at(uint16_t i) const { assert(i < _size); return _data.get()[i]; }
